Code is now compilable without debug flag

+++ srcs/env/parse_herited.c
@@ -6,7 +6,7 @@
 /*   By: jberredj <jberredj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/09 12:41:15 by jberredj          #+#    #+#             */
-/*   Updated: 2021/11/10 10:47:58 by jberredj         ###   ########.fr       */
+/*   Updated: 2021/11/10 13:43:44 by jberredj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	parse_herited_envp(t_env *env, char **envp)
 		nbr_var++;
 	env->nbr_vars = nbr_var;
 	i = -1;
-	while (++i < nbr_var)
+	while (++i < (int)nbr_var)
 	{
 		env_elem = create_env_var_from_str(envp[i], i); // TODO, PROTEGE LE CODE SALE MACAQUE
 		lst_elem = ft_lstnew(env_elem);
